add start function
