+++ src/deck.js
@@ -1,0 +1,15 @@
+const createDeck = (cards) => {
+
+    let deck = cards
+   return deck
+}
+
+const countDeck = (deck) => {
+
+   return deck.length
+}
+
+module.exports = {
+    createDeck,
+    countDeck
+    }
