+++ src/game.js
@@ -1,9 +1,18 @@
 const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
+const {createCard} = require('../src/card')
+const {evaluateGuess} = require('../src/card')
+const {createDeck} = require('../src/deck')
+const {countDeck} = require('../src/deck')
+const {createRound} = require('../src/round')
+const {takeTurn} = require('../src/turn')
+const {calculatePercentCorrect} = require('../src/round')
+const {endRound} = require('../src/round')
+
 
 function printMessage(deck) {
-  console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
+  console.log(`Welcome to FlashCards! You are playing with ${countDeck(deck)} cards.
   -----------------------------------------------------------------------`);
 }
 
@@ -11,4 +20,18 @@ function printQuestion(round) {
   util.main(round);
 }
 
-module.exports = { printMessage, printQuestion };
+function start() {
+  const questions = [];
+  prototypeQuestions.forEach(question => {
+    questions.push(createCard(question.id, question.question, question.answers, question.correctAnswer))
+  })
+  const deck = createDeck(questions)
+  const round = createRound(deck)
+
+  printMessage(deck)
+  printQuestion(round)
+}
+
+start()
+
+module.exports = { start,printMessage, printQuestion };
